+++ sdk/apimanagement/arm-apimanagement/src/apiManagementClientContext.ts
@@ -43,7 +43,7 @@ export class ApiManagementClientContext extends coreClient.ServiceClient {
       credential: credentials
     };
 
-    const packageDetails = `azsdk-js-arm-apimanagement/30.0.0-beta.1`;
+    const packageDetails = `azsdk-js-arm-apimanagement/8.0.0-beta.1`;
     const userAgentPrefix =
       options.userAgentOptions && options.userAgentOptions.userAgentPrefix
         ? `${options.userAgentOptions.userAgentPrefix} ${packageDetails}`