+++ sdk/identity/identity/src/index.ts
@@ -8,8 +8,19 @@ export { IdentityPlugin } from "./plugins/provider";
 import { TokenCredential } from "@azure/core-auth";
 import { DefaultAzureCredential } from "./credentials/defaultAzureCredential";
 
+export {
+  AuthenticationError,
+  ErrorResponse,
+  AggregateAuthenticationError,
+  AuthenticationErrorName,
+  AggregateAuthenticationErrorName,
+  CredentialUnavailableError,
+  CredentialUnavailableErrorName,
+  AuthenticationRequiredError,
+  AuthenticationRequiredErrorOptions
+} from "./errors";
+
 export { AuthenticationRecord } from "./msal/types";
-export { AuthenticationRequiredError } from "./msal/errors";
 export { serializeAuthenticationRecord, deserializeAuthenticationRecord } from "./msal/utils";
 export { TokenCredentialOptions } from "./client/identityClient";
 
@@ -53,8 +64,8 @@ export { AuthorizationCodeCredential } from "./credentials/authorizationCodeCred
 export { AzurePowerShellCredential } from "./credentials/azurePowerShellCredential";
 export { AzurePowerShellCredentialOptions } from "./credentials/azurePowerShellCredentialOptions";
 export {
-  AzureApplicationCredential as ApplicationCredential,
-  AzureApplicationCredentialOptions as ApplicationCredentialOptions
+  AzureApplicationCredential,
+  AzureApplicationCredentialOptions
 } from "./credentials/azureApplicationCredential";
 
 export {
@@ -70,16 +81,6 @@ export {
 } from "./credentials/onBehalfOfCredentialOptions";
 
 export { TokenCachePersistenceOptions } from "./msal/nodeFlows/tokenCachePersistenceOptions";
-
-export {
-  AuthenticationError,
-  ErrorResponse,
-  AggregateAuthenticationError,
-  AuthenticationErrorName,
-  AggregateAuthenticationErrorName,
-  CredentialUnavailableError,
-  CredentialUnavailableErrorName
-} from "./client/errors";
 
 export { TokenCredential, GetTokenOptions, AccessToken } from "@azure/core-auth";
 export { logger } from "./util/logging";